Add files via upload
By measuring the maximum Log-likelihood of the estimates of the GMM model using the proposed algorithms against other algorithms in the literature, we showed that the TM-based EM achieves more accurate estimates. The estimated models of the distributed algorithms were compared with respect to the central solution of the EM algorithm.